+++ CHANGELOG.md
@@ -3,7 +3,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
-### 🔄 Changed
+## StreamChat
+### 🐞 Fixed
+- Fix not able to mark channel read after clearing history [#2867](https://github.com/GetStream/stream-chat-swift/pull/2867)
+
+## StreamChatUI
+### 🐞 Fixed
+- Fix composer not interactable after enabling send-message capability [#2866](https://github.com/GetStream/stream-chat-swift/pull/2866)
 
 # [4.41.0](https://github.com/GetStream/stream-chat-swift/releases/tag/4.41.0)
 _November 03, 2023_